+++ src/queries/index.js
@@ -3,3 +3,4 @@ export * from "./getJobDetail";
 export * from "./postComment";
 export * from "./getInfiniteScrollItems";
 export * from "./getDummyApi";
+export * from "./getCommunityItems";
