+++ src/utils/axiosConfig.js
@@ -10,7 +10,7 @@ const axiosInstance = Axios.create({
 
 axiosInstance.interceptors.request.use((config) => {
   const token = localStorage.getItem("token");
-  if (token) {
+  if (token.accessToken) {
     config.headers.Authorization = `Bearer ${token}`;
   }
   return config;
